prerelease phase 1
this version brings object metadata caching, better tests, and more or less
feature completeness.

note: we're using our own home rolled assertion library. this is not best practice, but
it made sense as it was a simple file to implement and lets us add assertion types later.
we can also roll this back if we feel it's not a good look (i.e. switch back to testify/assert).